% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_sentinel.R
\name{aligner_sentinel}
\alias{aligner_sentinel}
\title{Aligner un raster Sentinel sur la grille de l'AOI}
\usage{
aligner_sentinel(sentinel, reference, target_res = 10)
}
\arguments{
\item{sentinel}{SpatRaster Sentinel (S1 ou S2)}

\item{reference}{SpatRaster de reference (ex: RGBI a 0.2m)}

\item{target_res}{Resolution cible en metres (defaut: 10)}
}
\value{
SpatRaster aligne sur l'emprise de reference
}
\description{
Reechantillonne un raster Sentinel (10m) pour qu'il couvre exactement
la meme emprise que le raster de reference, tout en gardant sa
resolution native (10m).
}
