% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flair_inference.R
\name{assembler_resultats_flair}
\alias{assembler_resultats_flair}
\title{Assembler les resultats FLAIR en GeoPackage et statistiques}
\usage{
assembler_resultats_flair(
  raster_classe,
  classes = NULL,
  dossier_sortie = "outputs"
)
}
\arguments{
\item{raster_classe}{SpatRaster mono-bande de classification}

\item{classes}{Table des classes (issue de \code{\link[=classes_cosia]{classes_cosia()}} ou \code{\link[=classes_lpis]{classes_lpis()}})}

\item{dossier_sortie}{Repertoire de sortie}
}
\value{
Liste avec \code{raster} et \code{stats}
}
\description{
Assembler les resultats FLAIR en GeoPackage et statistiques
}
