% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain.R
\name{calculer_derives_terrain}
\alias{calculer_derives_terrain}
\title{Calculer les derives morphologiques du MNT}
\usage{
calculer_derives_terrain(dtm)
}
\arguments{
\item{dtm}{SpatRaster mono-bande (DTM/MNT a 1m)}
}
\value{
Liste nommee de SpatRasters mono-bande :
\code{SLOPE}, \code{ASPECT}, \code{TPI}, \code{TWI}
}
\description{
A partir d'un DTM (Modele Numerique de Terrain) a 1m de resolution,
calcule les derives topographiques utiles pour la segmentation forestiere :
pente, orientation (aspect), TPI (Topographic Position Index) et
TWI (Topographic Wetness Index).
}
