% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flair_classes.R
\name{classes_cosia}
\alias{classes_cosia}
\title{Classes d'occupation du sol CoSIA (15 classes actives + non classifie)}
\usage{
classes_cosia()
}
\value{
Un data.frame avec les colonnes code, classe, couleur
}
\description{
Table des classes d'occupation du sol CoSIA apres remapping
depuis les codes FLAIR-1. Les classes desactivees (coupe, mixte,
ligneux, autre) sont remappees a 0 (non classifie).
}
\details{
Compatible avec la sortie des modeles 19 classes apres remapping.
}
\examples{
cls <- classes_cosia()
cls[cls$code <= 5, ]
}
