% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flair_classes.R
\name{classes_cosia_15}
\alias{classes_cosia_15}
\title{Classes CoSIA pour modeles FLAIR-INC 15 classes}
\usage{
classes_cosia_15()
}
\value{
Un data.frame avec les colonnes code, classe, couleur
}
\description{
Table des 15 classes utilisees par les modeles FLAIR-INC,
avec codes CoSIA 1-indexed (apres remapping Python).
Compatible avec la sortie de \code{predire_raster_complet()}.
}
