% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flair_classes.R
\name{classes_lpis}
\alias{classes_lpis}
\title{Classes de cultures LPIS/RPG (23 classes)}
\usage{
classes_lpis()
}
\value{
Un data.frame avec les colonnes code, classe, couleur
}
\description{
Table des 23 classes de cultures du jeu de donnees FLAIR-HUB,
issues du Registre Parcellaire Graphique (RPG / LPIS).
}
\examples{
cls <- classes_lpis()
head(cls)
}
