% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{combine_rgbi_mnt}
\alias{combine_rgbi_mnt}
\title{Combiner RGBI + MNT en image 5 bandes}
\usage{
combine_rgbi_mnt(rgbi, mnt)
}
\arguments{
\item{rgbi}{SpatRaster 4 bandes (Rouge, Vert, Bleu, PIR)}

\item{mnt}{SpatRaster 1 bande (MNT)}
}
\value{
SpatRaster 5 bandes (Rouge, Vert, Bleu, PIR, MNT)
}
\description{
Ajoute le MNT comme 5eme bande a l'image RGBI. Le MNT est reechantillonne
sur la grille RGBI si necessaire.
}
