% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{combine_rvb_irc}
\alias{combine_rvb_irc}
\title{Combiner les ortho RVB et IRC en image 4 bandes RGBI}
\usage{
combine_rvb_irc(rvb, irc)
}
\arguments{
\item{rvb}{SpatRaster ortho RVB (3 bandes : Rouge, Vert, Bleu)}

\item{irc}{SpatRaster ortho IRC (3 bandes : PIR, Rouge, Vert)}
}
\value{
SpatRaster 4 bandes (Rouge, Vert, Bleu, PIR)
}
\description{
Les modeles MAESTRO/FLAIR attendent 4 canaux optiques : Rouge, Vert, Bleu,
PIR. L'IRC IGN fournit le PIR en premiere bande.
}
