% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_indices.R
\name{compute_gndvi}
\alias{compute_gndvi}
\title{Calculer le GNDVI (Green Normalized Difference Vegetation Index)}
\usage{
compute_gndvi(rgbi, nir_band = 4L, green_band = 2L)
}
\arguments{
\item{rgbi}{SpatRaster avec bandes R, G, B, NIR}

\item{nir_band}{Indice de la bande PIR (defaut: 4)}

\item{green_band}{Indice de la bande Verte (defaut: 2)}
}
\value{
SpatRaster mono-bande avec valeurs GNDVI entre -1 et 1
}
\description{
GNDVI = (PIR - Vert) / (PIR + Vert)
}
