% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_indices.R
\name{compute_ndvi}
\alias{compute_ndvi}
\title{Calculer le NDVI (Normalized Difference Vegetation Index)}
\usage{
compute_ndvi(rgbi, nir_band = 4L, red_band = 1L)
}
\arguments{
\item{rgbi}{SpatRaster avec bandes R, G, B, NIR (ou 2 bandes PIR, Rouge)}

\item{nir_band}{Indice de la bande PIR (defaut: 4 pour RGBI)}

\item{red_band}{Indice de la bande Rouge (defaut: 1 pour RGBI)}
}
\value{
SpatRaster mono-bande avec valeurs NDVI \link{-1, 1}
}
\description{
Calcule le NDVI a partir d'un raster RGBI ou de bandes PIR et Rouge separees.
NDVI = (PIR - Rouge) / (PIR + Rouge)
}
\examples{
\dontrun{
rgbi <- terra::rast("ortho_rgbi.tif")
ndvi <- compute_ndvi(rgbi)
terra::plot(ndvi)
}
}
