% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_indices.R
\name{compute_savi}
\alias{compute_savi}
\title{Calculer le SAVI (Soil Adjusted Vegetation Index)}
\usage{
compute_savi(rgbi, nir_band = 4L, red_band = 1L, L = 0.5)
}
\arguments{
\item{rgbi}{SpatRaster avec bandes R, G, B, NIR}

\item{nir_band}{Indice de la bande PIR (defaut: 4)}

\item{red_band}{Indice de la bande Rouge (defaut: 1)}

\item{L}{Facteur d'ajustement du sol (defaut: 0.5)}
}
\value{
SpatRaster mono-bande avec valeurs SAVI
}
\description{
SAVI = ((PIR - Rouge) / (PIR + Rouge + L)) * (1 + L)
}
