% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\docType{data}
\name{CONDA_ENV}
\alias{CONDA_ENV}
\title{Nom de l'environnement conda pour MAESTRO}
\format{
An object of class \code{character} of length 1.
}
\usage{
CONDA_ENV
}
\description{
Nom de l'environnement conda pour MAESTRO
}
\keyword{datasets}
