% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flair_analysis.R
\name{cross_landcover_dem}
\alias{cross_landcover_dem}
\title{Correler l'occupation du sol avec le DEM}
\usage{
cross_landcover_dem(raster_classe, dem, classes = NULL)
}
\arguments{
\item{raster_classe}{SpatRaster de classification}

\item{dem}{SpatRaster du Modele Numerique de Terrain}

\item{classes}{Table des classes}
}
\value{
data.frame avec altitude moyenne, min, max par classe
}
\description{
Calcule les statistiques d'altitude par classe d'occupation du sol.
}
