% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flair_download.R
\name{.flair_modality_mapping}
\alias{.flair_modality_mapping}
\title{Mapping des noms de modalites vers les suffixes ZIP FLAIR-HUB}
\usage{
.flair_modality_mapping()
}
\value{
Liste nommee : nom interne -> suffixe dans le nom du ZIP
}
\description{
Mapping des noms de modalites vers les suffixes ZIP FLAIR-HUB
}
\keyword{internal}
