% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ign.R
\name{.lidar_coverage_pct}
\alias{.lidar_coverage_pct}
\title{Mesurer la couverture LiDAR HD (DSM) d'un raster WMS}
\usage{
.lidar_coverage_pct(r)
}
\arguments{
\item{r}{SpatRaster a evaluer (1 bande)}
}
\value{
Pourcentage \verb{[0, 100]} de pixels finis et non nuls
}
\description{
Le WMS LiDAR HD renvoie un raster valide y compris hors couverture
(pixels NA ou nuls). Cette fonction quantifie le \% de pixels effectifs.
}
\keyword{internal}
