% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finetune.R
\name{.treesatai_check_structure}
\alias{.treesatai_check_structure}
\title{Verifier la structure du dataset TreeSatAI}
\usage{
.treesatai_check_structure(data_dir, modality = "aerial")
}
\description{
Verifier la structure du dataset TreeSatAI
}
\keyword{internal}
