% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finetune.R
\name{.treesatai_download_one}
\alias{.treesatai_download_one}
\title{Telecharger un fichier depuis Zenodo}
\usage{
.treesatai_download_one(zenodo_base, fname, output_dir, mod)
}
\description{
Telecharger un fichier depuis Zenodo
}
\keyword{internal}
