% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finetune.R
\name{.treesatai_download_zenodo}
\alias{.treesatai_download_zenodo}
\title{Telecharger TreeSatAI depuis Zenodo (per-espece)}
\usage{
.treesatai_download_zenodo(output_dir, modalities)
}
\description{
Telecharger TreeSatAI depuis Zenodo (per-espece)
}
\keyword{internal}
