% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finetune.R
\name{.treesatai_find_tif_root}
\alias{.treesatai_find_tif_root}
\title{Trouver le dossier racine contenant les .tif apres extraction}
\usage{
.treesatai_find_tif_root(extract_dir, modality)
}
\description{
Trouver le dossier racine contenant les .tif apres extraction
}
\keyword{internal}
