% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finetune.R
\name{.treesatai_get_split_info}
\alias{.treesatai_get_split_info}
\title{Obtenir les informations de split train/test depuis split.zip}
\usage{
.treesatai_get_split_info(snapshot_path, output_dir)
}
\description{
Obtenir les informations de split train/test depuis split.zip
}
\keyword{internal}
