% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finetune.R
\name{.treesatai_move_species}
\alias{.treesatai_move_species}
\title{Deplacer les dossiers d'especes d'un split (move au lieu de copy)}
\usage{
.treesatai_move_species(src_split, dst_split, species)
}
\description{
Deplacer les dossiers d'especes d'un split (move au lieu de copy)
}
\keyword{internal}
