% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finetune.R
\name{.treesatai_resolve_hf_file}
\alias{.treesatai_resolve_hf_file}
\title{Resoudre un fichier HF (gere les symlinks casses sur Windows)}
\usage{
.treesatai_resolve_hf_file(snapshot_path, filename)
}
\description{
Sur Windows sans privileges admin, hfhub cree des symlinks qui echouent.
On lit le pointer file pour trouver le blob reel dans le cache HF.
}
\keyword{internal}
