% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/essences.R
\name{essences_pureforest}
\alias{essences_pureforest}
\title{Classes d'essences forestieres PureForest}
\usage{
essences_pureforest()
}
\value{
Un data.frame avec les colonnes code, classe, nom_latin et type
}
\description{
Table des 13 classes d'essences forestieres du jeu de donnees PureForest
(IGN), utilise pour l'entrainement du modele MAESTRO. Source officielle :
fiche dataset Hugging Face \code{IGNF/PureForest}.
}
\examples{
ess <- essences_pureforest()
ess[ess$type == "feuillu", ]
}
