% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flair_analysis.R
\name{export_raster}
\alias{export_raster}
\title{Exporter un raster de classification}
\usage{
export_raster(r, output_path)
}
\arguments{
\item{r}{SpatRaster a exporter}

\item{output_path}{Chemin du fichier de sortie}
}
\value{
Chemin du fichier (invisible)
}
\description{
Sauvegarde un raster avec compression LZW.
}
