% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patches.R
\name{extraire_patches_multimodal}
\alias{extraire_patches_multimodal}
\title{Extraire les patches multi-modaux depuis plusieurs SpatRasters}
\usage{
extraire_patches_multimodal(modalites, grille, taille_pixels = 250)
}
\arguments{
\item{modalites}{Liste nommee de SpatRasters (ex: \code{list(aerial=..., dem=..., s2=...)})}

\item{grille}{sf grille de patches (issue de \code{\link[=creer_grille_patches]{creer_grille_patches()}})}

\item{taille_pixels}{Taille cible pour aerial/dem en pixels (defaut: 250)}
}
\value{
Liste de listes nommees, chaque element contient les matrices
(H*W x C) pour chaque modalite. Ex: \code{patches[[i]]$aerial}, \code{patches[[i]]$s2}
}
\description{
Pour chaque patch de la grille, extrait les valeurs de chaque modalite
(aerial, dem, s2, s1_asc, s1_des) separement. Chaque modalite est
reechantillonnee a sa taille de patch native :
\itemize{
\item aerial : 250 x 250 pixels (0.2m)
\item dem    : 50 x 50 pixels   (1m)
\item s2, s1_asc, s1_des : 5 x 5 pixels (10m)
}
}
\details{
Les donnees sont structurees pour etre passees au modele MAESTRO.
}
