% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{find_checkpoint_name}
\alias{find_checkpoint_name}
\title{Trouver le nom du fichier checkpoint dans un depot Hugging Face}
\usage{
find_checkpoint_name(hf_repo)
}
\arguments{
\item{hf_repo}{Identifiant du depot (ex: \code{"IGNF/MAESTRO_FLAIR-HUB_base"})}
}
\value{
Nom du fichier checkpoint, ou NULL
}
\description{
Interroge l'API Hugging Face pour lister les fichiers du depot et
retourne le fichier de poids le plus probable.
}
