% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flair_classes.R
\name{flair_models}
\alias{flair_models}
\title{Modeles FLAIR-HUB pre-entraines disponibles}
\usage{
flair_models()
}
\value{
Un data.frame avec les colonnes id, architecture, encoder, decoder,
n_bands, supervision, miou
}
\description{
Catalogue des modeles pre-entraines pour la segmentation d'occupation du sol.
Les modeles sont heberges sur HuggingFace par l'IGNF.
}
\examples{
mods <- flair_models()
mods[mods$miou > 60, ]
}
