% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flair_inference.R
\name{flair_python_module_path}
\alias{flair_python_module_path}
\title{Chemin vers le module Python d'inference FLAIR}
\usage{
flair_python_module_path()
}
\value{
Chemin du repertoire Python (character)
}
\description{
Chemin vers le module Python d'inference FLAIR
}
