% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ign.R
\name{ign_layer_name}
\alias{ign_layer_name}
\title{Construire le nom de couche WMS IGN selon le millesime}
\usage{
ign_layer_name(type = c("ortho", "irc"), millesime = NULL)
}
\arguments{
\item{type}{\code{"ortho"} ou \code{"irc"}}

\item{millesime}{\code{NULL} (mosaique la plus recente) ou entier (ex: 2023)}
}
\value{
Nom de couche WMS (character)
}
\description{
Construire le nom de couche WMS IGN selon le millesime
}
\examples{
ign_layer_name("ortho")
ign_layer_name("irc", 2023)
}
