% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoi.R
\name{load_aoi}
\alias{load_aoi}
\title{Charger une zone d'interet depuis un GeoPackage}
\usage{
load_aoi(gpkg_path, layer = NULL)
}
\arguments{
\item{gpkg_path}{Chemin vers le fichier .gpkg}

\item{layer}{Nom de la couche (NULL = premiere couche)}
}
\value{
sf object en Lambert-93 (EPSG:2154)
}
\description{
Lit le fichier GeoPackage et reprojette en Lambert-93 (EPSG:2154)
si necessaire. Affiche un resume de l'emprise et de la surface.
}
\examples{
\dontrun{
aoi <- load_aoi("ma_zone.gpkg")
}
}
