% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flair_download.R
\name{load_flair_patch}
\alias{load_flair_patch}
\title{Charger un patch FLAIR-HUB depuis un fichier GeoTIFF}
\usage{
load_flair_patch(tif_path)
}
\arguments{
\item{tif_path}{Chemin vers le fichier GeoTIFF du patch}
}
\value{
SpatRaster
}
\description{
Charger un patch FLAIR-HUB depuis un fichier GeoTIFF
}
