% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_bdforet.R
\name{mapping_tfv_ndp0}
\alias{mapping_tfv_ndp0}
\title{Mapping des codes TFV vers les classes NDP0}
\usage{
mapping_tfv_ndp0()
}
\value{
data.frame avec colonnes tfv_pattern (regex) et code_ndp0
}
\description{
Retourne un data.frame de correspondance entre les codes TFV de la
BD Foret V2 et les 10 classes NDP0.
}
