% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_indices.R
\name{mask_vegetation}
\alias{mask_vegetation}
\title{Creer un masque de vegetation par seuillage NDVI}
\usage{
mask_vegetation(rgbi, seuil = 0.3, nir_band = 4L, red_band = 1L)
}
\arguments{
\item{rgbi}{SpatRaster avec bandes R, G, B, NIR}

\item{seuil}{Seuil NDVI pour la vegetation (defaut: 0.3)}

\item{nir_band}{Indice de la bande PIR (defaut: 4)}

\item{red_band}{Indice de la bande Rouge (defaut: 1)}
}
\value{
SpatRaster binaire (1 = vegetation, 0 = non-vegetation)
}
\description{
Creer un masque de vegetation par seuillage NDVI
}
