% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{python_module_path}
\alias{python_module_path}
\title{Chemin vers le module Python d'inference MAESTRO}
\usage{
python_module_path()
}
\value{
Chemin du repertoire Python (character)
}
\description{
Retourne le chemin du repertoire contenant \code{maestro_inference.py},
installe avec le package dans \verb{inst/python/}.
}
