% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_sentinel.R
\name{s2_bands_config}
\alias{s2_bands_config}
\title{Configuration des bandes Sentinel-2}
\usage{
s2_bands_config()
}
\value{
Liste nommee avec nom, longueur d'onde et resolution par bande
}
\description{
Les 10 bandes spectrales utilisees par MAESTRO pour la modalite S2.
}
