% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patches.R
\name{taille_patch_modalite}
\alias{taille_patch_modalite}
\title{Taille de patch par modalite MAESTRO}
\usage{
taille_patch_modalite(mod_name, taille_pixels_ref = 256L)
}
\arguments{
\item{mod_name}{Nom de la modalite}

\item{taille_pixels_ref}{Taille pour aerial/dem si mod_name n'est pas connue}
}
\value{
Entier : nombre de pixels du patch
}
\description{
Wrapper retro-compatible autour de \code{\link[=modalite_specs]{modalite_specs()}}. Pour de nouveaux
developpements, preferer l'acces direct via \verb{modalite_specs()$<mod>$image_size}.
}
