% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_bdforet.R
\name{tfv_to_ndp0}
\alias{tfv_to_ndp0}
\title{Convertir un code TFV en classe NDP0}
\usage{
tfv_to_ndp0(code_tfv)
}
\arguments{
\item{code_tfv}{Vecteur de codes TFV (character)}
}
\value{
Vecteur d'entiers (codes NDP0, 9 = non-foret par defaut)
}
\description{
Convertir un code TFV en classe NDP0
}
