% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/essences.R
\name{treesatai_species}
\alias{treesatai_species}
\title{Les 20 genres d'arbres du dataset TreeSatAI}
\usage{
treesatai_species()
}
\value{
Vecteur character de 20 noms de genres
}
\description{
Retourne le vecteur des 20 genres d'arbres europeens presents dans
le TreeSatAI Benchmark Archive. Utilise en interne pour le telechargement
et l'organisation des donnees.
}
