% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ign.R
\name{validate_wms_data}
\alias{validate_wms_data}
\title{Verifier qu'un raster WMS contient des donnees reelles}
\usage{
validate_wms_data(r, min_pct = 5)
}
\arguments{
\item{r}{SpatRaster a valider}

\item{min_pct}{Pourcentage minimum de pixels non-vides requis (defaut: 5)}
}
\value{
TRUE si le raster contient suffisamment de donnees
}
\description{
Certaines couches millésimees ne couvrent pas toutes les zones. Le WMS
retourne alors un raster valide mais vide (pixels a 0 ou NA).
}
