% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flair_classes.R
\name{classes_cosia}
\alias{classes_cosia}
\title{Classes d'occupation du sol CoSIA (19 classes)}
\usage{
classes_cosia()
}
\value{
Un data.frame avec les colonnes code, classe, couleur
}
\description{
Table des 19 classes d'occupation du sol du jeu de donnees FLAIR-HUB,
issue de la photo-interpretation experte CoSIA (IGN).
}
\examples{
cls <- classes_cosia()
cls[cls$code <= 5, ]
}
