% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patches.R
\name{creer_grille_patches}
\alias{creer_grille_patches}
\title{Creer une grille de patches pour l'inference}
\usage{
creer_grille_patches(aoi, taille_patch_m = 50)
}
\arguments{
\item{aoi}{sf object en Lambert-93}

\item{taille_patch_m}{Taille des patches en metres (defaut: 50)}
}
\value{
sf data.frame (grille de patches avec colonne \code{id})
}
\description{
Genere une grille reguliere de patches carres couvrant l'AOI.
Seuls les patches qui intersectent l'AOI sont conserves.
}
