% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patches.R
\name{taille_patch_modalite}
\alias{taille_patch_modalite}
\title{Taille de patch par modalite MAESTRO}
\usage{
taille_patch_modalite(mod_name, taille_pixels_ref = 250L)
}
\arguments{
\item{mod_name}{Nom de la modalite}

\item{taille_pixels_ref}{Taille de reference (defaut: 250 pour aerial)}
}
\value{
Entier : nombre de pixels pour ce patch
}
\description{
Retourne la taille de patch en pixels attendue par le modele MAESTRO
pour chaque modalite, selon la resolution native :
\itemize{
\item aerial : 250 pixels (0.2m, 50m / 0.2m)
\item dem    : 50 pixels  (1m, 50m / 1m)
\item s2, s1 : 5 pixels   (10m, 50m / 10m)
}
}
